requirements.txt
# Sentinel-2 Water Quality Processing Requirements

# Core scientific computing packages
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0
netcdf4>=1.6.0

# Geospatial processing packages
geopandas>=0.14.0
shapely>=2.0.0
cartopy>=0.21.0
rasterio>=1.3.0

# Visualization packages
matplotlib>=3.7.0
seaborn>=0.12.0
cmocean>=3.0.0

# HTTP and API packages
requests>=2.31.0
urllib3>=2.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dateutil>=2.8.0

# Optional: Enhanced plotting and analysis
plotly>=5.15.0
folium>=0.14.0

# Optional: Additional geospatial tools
pyproj>=3.6.0
fiona>=1.9.0
